Extract configuration static inner classes to separate classes
